The Origin Behind My Love For The Number 27
I don't really know why the number 27 is my favorite number. The week I had to create my first ever username, "Halil" was taken by another user. It suggested that I add some numbers add the end. The suggestions were some weird numbers. I recalled an ad I saw about TheLegend27 advertising some mobile game. It was everywhere on Youtube at the time and it was stuck in my head. That's why I decided to set my username to "Halil27". But that was taken as well. So, I added another 27 and that was good enough. From then on, I set "Halil2727" or some variation of it as my username on literally everything. That's the origin story behind my username. Pretty cool, right?
I actually need some indication that you guys are reading this so when you commit your next change, put something when it says to add an optional description.