Can't Check Input For Some Situations--- Read Desc
First of all, Willy Wonka's situations work perfectly now. It will not accept any input other than 1, 2, or 3 (or yes or no in the "do you take this seriously?" question). You should still check these to see if I missed anything but they should work perfectly. Ok, so the problem is, the method I used to achieve this does not work for "attractive" situations. Because of the way "attractive" options are structured (they print the next situation directly from a void method in another class), it is impossible to check the input on those using the method I created. If you can make a method for those options, great. But I don't think it is possible without changing the methods, drastically (which at this stage might do more harm than good). So, leave it as is.